Don't clobber methods that already exist.
This change prevents clobbering instance methods that already exist on the
MBeanProxy class. I ran into it when working with an mbean that exposed
an 'initialize' operation, which made it impossible to instantiate an
instance of the MBeanProxy subclass for that mbean a second time.

A side-effect of this change is you'll end up with a bunch of 'mbean_*'
methods defined after proxying a two mbeans of the same class.